Users Registration finished

+++ app/controllers/panel/users_controller.rb
@@ -1,23 +1,24 @@
 class Panel::UsersController < BaseController
   before_action :set_user, only: [:edit, :update, :destroy]
-  #after_action :verify_authorized, only: [:new, :destroy]
-  #after_action :verify_policy_scoped, only: :index
+  after_action :verify_authorized, only: [:new, :destroy]
+  after_action :verify_policy_scoped, only: :index
 
   def index
-  	@users = User.all
+  	#@users = User.all
+    #@users = User.with_full_access
     #@users = User.with_restricted_access
-    #@users = policy_scope(User)
+    @users = policy_scope(User)
   end
 
   def new
   	@user = User.new
-    #authorize @user
+    authorize @user
   end
 
   def create
-  	@user = User.new(user)
+    @user = User.new(params_user)
     if @user.save
-      redirect_to backoffice_users_path, notice: "El Usuario (#{@user.email}) fue creado con exito!"
+      redirect_to panel_users_path, notice: "El Usuario (#{@user.email}) fue creado con exito!"
     else
       render :new
     end
@@ -29,19 +30,19 @@ class Panel::UsersController < BaseController
 
   def update
     if @user.update(params_user)
-      UserMailer.update_email(current_user, @user).deliver_now
-      redirect_to backoffice_users_path, notice: "El Usuario (#{@user.email}) fue actualizado con exito!"
+      #UserMailer.update_email(current_user, @user).deliver_now
+      redirect_to panel_users_path, notice: "El Usuario (#{@user.email}) fue actualizado con exito!"
     else
       render :edit
     end
   end
 
   def destroy
-    #authorize @user
-    #user_name = @user.user_name
+    authorize @user
+    user_name = @user.name
 
     if @user.destroy
-      redirect_to backoffice_users_path, notice: "El Usuario (#{user_email}) fue excluido con exito!"
+      redirect_to panel_users_path #notice: "El Usuario (#{user_email}) fue excluido con exito!"
     else
       render :index
     end
@@ -49,17 +50,24 @@ class Panel::UsersController < BaseController
 
   private
 
-  def set_user
-    @user = User.find(params[:id])
-  end
-
-  def params_user
-    #passwd = params[:user][:password]
-    #passwd_confirmation =  params[:user][:password_confirmation]
-
-    if passwd.blank? && passwd_confirmation.blank?
-    #  params[:user].except!(:password, :password_confirmation)
+    def set_user
+      @user = User.find(params[:id])
     end
-    #params.require(:user).permit(policy(@user).permitted_attributes)
-  end
+
+    def params_user
+      if password_blank?
+        params[:user].except!(:password, :password_confirmation)
+      end
+
+      if @admin.blank?
+        params.require(:user).permit(:username, :name, :email, :role, :password, :password_confirmation)
+      else
+        params.require(:user).permit(policy(@user).permitted_attributes)
+      end
+    end
+
+    def password_blank?
+      params[:user][:password].blank? &&
+      params[:user][:password_confirmation].blank?
+    end
 end
